third_party/blink/web_tests/platform/generic/fast/forms/datetimelocal-multiple-fields/datetimelocal-multiple-fields-reset-value-after-reloads-expected.txt

Checks reloading page resets input values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Initial values:
PASS testInput1.value is ""
PASS testInput2.value is "1234-05-06T07:08"
Reloaded values:
PASS testInput1.value is ""
PASS testInput2.value is "1234-05-06T07:08"
PASS successfullyParsed is true

TEST COMPLETE
